Move all components with input fields specific to different quote types from Offer.tsx into one file each. Also move their initial form values there. Add getCurrentAvailableQuoteData function to get the right initialValues for the rendered form (initial form values are added as properties on the quotesByMarkets objects). Create and import some more types for this to work.
